API building & Testing
I added some more content to the different API endpoints.

The messages returned are a little more informative, and the functions are re-ordered.

`geopolitical` now has `geopoliticalunits` and an `id` element.

I removed all the original puppy stuff from the tutorial I was using, and have added in a `site` element, but this is not really implemented yet.