Added a semi-colon to the `sitebygpid` sql file.
Working on the taxon calls.  `taxonquery` works, but the `index` file is incomplete, but I've made changes to the variable outputs.

Cleaned out the `taxa.js` file so that the taxon queries are embedded in ther own SQL files.

There's one with a recursive and one without for `lower===true` condition.