There is a new parameter function that check body and url.
This helps manage data input for long strings, for example locations. Need to write tests for the method, but it looks like it's working generally.
Updated the sites methods as well to accept a few more parameters, and to work with the routing, where now everything can be passed into sitesquery.