gitflow-feature-stash: postdownloads
This looks long, but it is mostly just rewrites of the tests. Also added a flag in the genoatt.sh script to support testing on the remove dev server.
The big thing is sorting out how to manage POST/GET verbs so we can begin working with larger spatial objects for our filtering.