Adding tests and modifying the spatial query.
Working on the update the API to work better with spatial queries. Replaced Intercept with !Disjoint so we can account for intercepts but also contains/within queries. This helps manage polygon data better.
There are the beginnings of streaming elements in the set of functions, but they are not directly included here.