Updated the common functions in src/neotomaapi
One of the main issues is properly managing the parsing of WKT and geoJSON objects. We wanted to get a clean try/catch pathway to parse WKT and geoJSON objects. This was accomplished, but subsequently we needed to also be able to parse geoJSON objects more complex that terraformer allows. This resulted in the inclusion of the `customFilter()` function, which looks specifically for a field name within a complex JSON object, allowing recursive searching of the object.