Large refactoring of functions to make use of src/neotomaapi.js
I've moved a lot of code into src/neotomaapi.js, this now cleans out code in each of the subsequent helpers folders, and leads to a single point for modifying code.
Seems to work nicely. This also means that a lot of the main functions (sites, datasets, etc) can now be run from a single function, rather than using a sitesbydsid, sitesbyid, etcetera.  Thus, changes to the handler/routes.