Updated all calls to the proper db connection
The use of the global connection string also allowed some general code cleanup, particularly in older v1.5 calls.
Because of the change in the db connection we also needed to pass the `req` object into the checkObject() call,  since it is intended to explicitly check against existing vocabularies/terms, and so needs access to the db object (in req.local).
Reran tests and everything generlly looks good now. I need to adjust timeouts and also add a rate limiter.