Adding rate limiter.
I also removed a file in the databases folder.
The rate limiter uses express-rate-limit middleware.  I'm limiting at 50 calls every 10 seconds. That seems pretty fast. . . We can use another database to do fancier limiting, but for now I'm going to leave it as is.
We also skip the limit if a local environment variable is set (to allow local testing).