gitflow-feature-stash: icesheets
Creates a SQL query to match inputs (calibrated age, projection and precision) to the nearest ice sheet layer based on the existing Dyke layers in the `ap.icesheets` schema.
This query uses ST_Union() to merge sheets, removing the need to generate jpeg files for rendering in Explorer.
The function also includes the use of ST_MakeValid() because it appears that some of the icesheet layers include unclosed polygons. I'm not sure why.
The CTE allows us to query by ages that are not explicitly in the layers, by picking the ice sheet layer that is closest to the submitted calibrated age.