Working on Speleothem tests
The Speleothem endpoint was added, but needed some minor tweaks to make it pass the tests.

First, I added a new 'collectionunitidPath` to the openapi parameters. This allows us to test the endpoint properly.

To support collectionunitids or other parameters elsewhere, I used the `paramgrab` model in the speleothems.js file, along with `ifUndef()`.

This then captures the parameter in the path and passes it along cleanly to the query. The function seems to resolve now.