Working to get a NAT up and better error checking.
It looks like the issue is more likely a result of the use of the private subnet to deploy the app.
We have to use it to connect to the database, but that means that the app cant call out to external services. By setting up the NAT in the infrastructure.yaml we can then connect to other APIs.